add mssql collector take 1
We are transitioning from Bosun/scollector to Prometheus to monitor our
systems.

the scollector was collecting some mssql metrics and we didn't want to
lose that.  So this first attempt was to replicate that process.

I studied what scollector was doing to gather mssql metrics
(https://github.com/bosun-monitor/bosun/blob/master/cmd/scollector/collectors/sql_windows.go)
and replicated that in a new `mssql` collector.

It works, but there are some things that rub me the wrong way:

* I used the collector-generator tool to template 8 WMI classes, then
did a lot of cut-and-pasting to merge them into a single collector.

* the metrics that were generated were based on SQL Server 2016.  I
suspect that older version won't have the same WMI fields and may not
work.

* our servers each only have one instance using the default
"MSSQLSERVER" instance name.  this will not work on servers with differnet/multiple
names.

I can't help but think that there's a better way.  One thing I'm
thinking:

* Given a list of WMI class suffixes:  ['SQLServerGeneralStatistics', 'SQLServerLocks', ...]
* query WMI for a list of classes that match that `select * from Meta_Class WHERE __Class LIKE "Win32_PerfRawData_%_{suffix}`
* for each class, dynamically create list of metrics and collect them (still requires PoC)

the dynamic nature of that will make the code shorter (less cut-n-pasty)
and should work with multiple versions of windows as we will only be
generating metrics for classes/values that that instance of windows
knows about