favicon return 404 not found
during testing I found that the collectors were getting triggered twice
when hitting the exporter from chome or firefox.

turns out, they were trying to get `/favicon.ico` and the code was
redirecting that request to `/metrics`

added a explicit handler to return 404 NotFound when `/favicon.ico` is
requested.

This issue would not have been triggered with normal requests from prometheus,
only when testing with a browser.