mssql skip subsequent invalid class
Created a map struct to keep track of any any instance/class name pairs that generate an "Exception occurred. (Invalid class)" error.

First time an instance/class pair gets that error, it will add to that map.

Before querying WMI data, it will check if the current instance/class is in that map and, if so, skip querying WMI data.

So what will happen is that on the first scrape, the `Exception occurred. (Invalid class)` error will get logged, but subsequent scrapes will not attempt to query that instance/class anymore.  (Until the wmi_exporter service is restarted.)