Add initial HTTPS support for windows_exporter
HTTPS support is required for environments with policy and/or regulatory
requirements which require internal traffic to be encrypted.

This has been rather liberally taken from the `node_exporter` project,
with some minor modifications made to support the current
windows_exporter logging library and Windows error messages.

Addition of the code and test files will increase maintenance overhead
as we'll need to keep an eye on HTTP changes in `node_exporter`.
Unfortunately due to differences in platforms and logging libraries the
package couldn't be directly imported.

Signed-off-by: Ben Reedy <breed808@breed808.com>