docs: add actual Windows Update service name
To check if the _Windows Update_ service is running it is useful to run
the `Get-Service` cmdlet, therefore it's good to know the actual service
_name_ (as opposed to the _display name_ that is used e.g. by the
Services snap-in).

For example

```
Get-Service wuauserv
```

will show the current status of the update service.

Additionally, this commit modifies several *s* to be lowercase, as the
full *display name* of the corresponding service is "Windows Update",
not "Windows Update Service" - at least on all the Windows versions I
was able to check.

Signed-off-by: Niko Ehrenfeuchter <mail@he1ix.org>